pyproject.toml
[project]
name = "LicenseGuard-agent"
version = "0.1.0"
description = "An AI agent responsible for providing insights to the frontend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.121.2",
    "langchain-mcp-adapters>=0.1.13",
    "python-dotenv>=1.2.1",
]
